feat: add command exists checking for all tools installation

+++ linux/installation/developer-packages/ubuntu/tools/go.sh
@@ -1,8 +1,15 @@
 #! /bin/bash
 
 GO_VERSION=1.24.5
-wget "https://go.dev/dl/go$GO_VERSION.linux-amd64.tar.gz"
-tar xfvz go$GO_VERSION.linux-amd64.tar.gz
-cp go/bin/* /usr/local/bin/
-echo "==== CLEAN UP ===="
-rm -drf go$GO_VERSION.linux-amd64.tar.gz go
+
+if ! command -v go 2>&1 >/dev/null
+then
+    echo "[INSTALLING ⬇️ ] Go"
+    wget "https://go.dev/dl/go$GO_VERSION.linux-amd64.tar.gz"
+    tar xfvz go$GO_VERSION.linux-amd64.tar.gz
+    cp go/bin/* /usr/local/bin/
+    echo "==== CLEAN UP ===="
+    rm -drf go$GO_VERSION.linux-amd64.tar.gz go
+else
+    echo "- [CHECKED ✅] go command exists"
+fi
